Created a Loading Scene
Made a scene to plan out the loading screen. The loading screen would need to be spawned in, so you need to set up a prefab to be generated when a loading screen appears. You probably need to make a new layer as well, plus have it mute the audio or stop the game events or something.