Locked Doors During Tutorial
I locked the doors during the tutorial so that the player is limited in terms of what option they have at the start. There appears to be a glitch where the enemy for the second battle has more health than it should be allowed. This happens regardless of if the tutorial is active or not, so this seems to be a different sort of problem.