Finished Battle Text
Finished translating the battle text. The tutorial text still needs to be translated, and there is a glitch where the player tried to learn a move twice, so those are the next things to address. I also need to set up the battle entity randomization. I still haven't tried the evolution thing either.