Added in Tutorial Messages and Fixed Speed
Added in translated tutorial messages, and fixed the speed (it wasn't being gotten properly). Currently the game is around 25.4 MB, with the limit being 30 MB. I should add in the LOL connections and delete the template content so that I have a good idea of how big the game is. Also, the game didn't work when running it in the browser, so I need to fix that.