Working on Game Saving and Loading
You need to create a way to save and load the game. You'll likely create an object that gets brought into the title scene and then loads the data.

Also, you need to add colliders to read different on-screen prompts. You also still need to update the UI for showing the move energy usage.