Implemented Save and Continue Buttons
I implemented the save and continue buttons. I don't think I need the save content in the InitGame file, so I should remove that content. The continue button disappears if it can't be used, so you may need to adjust the start button's position so that it doesn't look awkward. Right now it automatically starts the game upon loading the data, but that may not be what you want.