Need to Set Up the Save System
I need to set up the save system for the game. I should be able to do some work tomorrow and finish that up quickly. As for learning moves, the player still gets recommended moves they already have, so that needs to be fixed.

Some things will need to be streamlined, so make sure to change the game design document.