Battle Bug Still There, Going to Leave it for Now
The battle bug is still in the game, but I'm going to leave it since it's not a big deal. I should fix it later though.

The bug where the player gets prompted twice to learn a move still seems to show the right textbox, it's just covered (assuming the TTS is correct). I don't know why this happens, and I need to figure that out. It seems to happen after the second battle won, but I'm not sure.