Couldn't Recreate the Move Learn Glitch, Leaving it for Now
I wasn't able to recreate the move learn glitch, so I'm going to leave it for now. The glitch appears to cause a learn prompt to trigger early (on the 'player won' message specifically), but I don't know why. It also seems to happen on the second battle, and may be effected by the tutorial, but I'm still not sure about that either.

Either way, I'm going to leave it for now. Hopefully it'll work itself out later, but I think I'm going to work on other parts of the game now.