Restore Enemy Health Regardless of Evolution + Save Info
I changed the game so that the enemies get full health and energy restores regardless of whether or not they evolved.

As for the save, it appears that the game still remembers your prior saves from testing. I'm not sure how to delete prior save data, but that's what appears to be happening.