Added Arrows for Textbox
I added arrows for the textbox instead of using labeled buttons for ease of use.

For the move learning, it appears to pop up before and after the player's level up. However, I don't know why this happens. I suppose a quick fix could be to have a flag for the player leveling up, and use that to determine if learning a new move is allowed, but I'm not sure how well that will work.