Fixed the Audio Issue for the Tutorial
Fixed an issue where the tutorial music wouldn't be played if the tutorial was running for the battle. I also translated the information on the init screen in case it shows to the user. When I ran it before the game ran poorly, but then I ran it again and it ran fine, so I'm not sure how well it's going to go.