Fixed Save Issue
Fixed the issue where the game was not saving properly. Turns out the text was not set and it was encountering a null error.

There's a bug with the music where the victory jingle overlaps with the BGM, and that the BGM gets slowed down when it's not supposed to be. Fix that.