Fixing the Save System for the Game.
Fixed the save system so that all the data is properly updated and implemented (the phase wasn't changing before). There's still a music overlap issue when the battle is going back to the overworld (namely when the jingle is involved), so that's an issue that still needs to be addressed.

The game currently sits at 26.8 MB.