Fixed the Stat Change Tutorial Bug
Turns out that the stat change tutorial was being called in the critical hit function as well. That has been fixed.

Also, the "attempt to learn a move twice" issue seems to also have been fixed naturally, so that's nice. I don't know what caused it before, but it seems alright now.