Added First Death Tutorial and Fixed First Turn Enemy Death During Tutorial
Added a tutorial prevents the player from dying during the first battle. I also made it so that the enemy can't die in one turn during the first battle in the tutorial. This is because it screws up the other tutorials if that happens. I haven't looked over all the moves yet, but the game seems pretty done at this point.