Changed Some Values, and Added Damage Animation
I changed some values for balance reasons. I realize that I may need to change the tutorial more so that the enemy is actually beatable.

I also added a damage animation for the enemy, which is just a quick red flash. I might add a yellow flash for status effects? I also need to add an animator for the move animations themselves.