Disabled the run button during the tutorial.
The run button gets disabled during the tutorial now. Also, I changed where the door sound effect is played so that it doesn't get repeated over and over. There appears to be a problem where the move isn't being reset properly during the tutorial for making the second move, so I need to briefly fix that.