Fixed the Tutorialization of Enemies
Fixed the tutorial so that a set of enemies are hand-picked so that the player can reasonably win. The early battle end doesn't appear to be working consistently, so that needs to be fixed. I also need to change how much health and energy get restored, because the game seems way too difficult right now. Maybe I should check and see if the player has an attacking move at all times?