Disabled Saving Before First Battle (tutorial).
I disabled saving the game before the first battle for the tutorial. This way there's no wonky saving that gets done.

There's a glitch where an index out of bounds error caused not all the pages to be shown during the boss fight, and the player ran away despite being dead. The game over happened the moment another battle began, but still. This needs to be addressed.