Stat Changes + Need to Fix Textbox Glitch
Made some small stat changes for game balance. Also, there appears to be an out of bounds error being encountered with the textbox, but I don't know why it's happening. I need to fix that. There's nothing on the provided line for index out of bounds exception, so I'm kind of confused on that.