Made Some Results Screen Fixes
Turns out the results screen was calculating the time wrong, so I just quickly fixed that. I should check how that information is being uploaded to the LOL system as well.

The game is completable, but it still has the move learning bug, and I don't now why it happens. The enemies also sometimes attack when they're dead. We found workarounds so that they don't break the game, but the information is still displayed, so I need to find a way to address it.