Fixed Tutorial Save Issue
Fixed the issue where the tutorial settings weren't saved to the saved game. I needed to do that so that it wouldn't bug out if the tutorial got turned on at a random point. The save system appears to work properly now. On a related note, if you stop the project and run it the MockLOL will remember a prior save, which is what you can use to test it.