Editor Settings
Pushing the editor settings. Closing the project didn't appear to get rid of the saved data, so it seems you need to stop the background processes or maybe even do a full computer reboot to test the system from having no data. You shouldn't need to test this, but you might as well to be safe.