Added Speak Message to Stop Closed Textbox from Being Read from Load
I added a speak message so that the loaded textbox isn't read. Stopping the text-to-speech didn't work, so I have an alternate message instead. There's no failed message, but that would show the tutorial textbox anyway, so eh.

Also, the LOL save data is retained even if you close the project and shut down the system. Well, at least it's good for testing.