Still Working on Things
Fixed a small error in the question saving system. The player gets asked a total of 5 questions, but the list only had a room for 4, so that's been fixed. Still not sure why the game doesn't work from the initialization phase anymore. Rebuilding the game was successful, but it's still causing problems when I try to play from the initialization phase, so I don't know what's happening.