Added Redoing Questions (WIP) and Changed Text Speed
I added the ability to redo questions, but I still need to make the question counter match up properly. Also, I changed the text-speed so that it doesn't function off of framerate. I really should find a better way to do the text speed, but tying it to framerate is a bad practice.