Current Progress
Pushing the current progress. I should note that there was a glitch with one of the new moves (I think Livewire?) where the button was enabled, but the move couldn't be performed. The energy level was the same as the energy needed for the move if memory serves, so that might be a rounding issue.