Fixed Continue Screen + Bug Acknowledgement
I fixed the continue screen so that it doesn't allow you replay a completed game anymore.

For some reason, the UFO MK2 doesn't upgrade to MK3. I assume it's because I reduced the total number of rooms, and that messed with something. It's the only enemy with three evolution stages, and it's not noticeable, so unless the game gets rejected I probably won't try to fix it.