Pushed Changes from Main
I pushed the changes from the main branch to the promo branch. I also changed the code stripping to go from high to medium. High and medium code stripping seem to leave the build size the same. The less code stripping, the better the game runs, so I should have the promo build be of a lesser code stripping amount.

The promo build size is 31.2 MB.