Changed Some Compression Settings + Other Small Edits
I changed some compression settings for the game, mainly for the audio. It is currently 13.3 MB in size.  

I also have the enemies that can't evolve get a level up during a phase change now, and adjusted the health and energy restores on game over. I should speed up the auto skip for the battle textbox, and see if I can have the back button disabled when on the first page of the tutorial textbox.