Text Scaler + Spanish Translation Successful
The Spanish Translation appears to work properly, so I won't need to submit a new build.

I also added a text scaler for shrinking text if t goes beyond the character limit. It's not being used at all, but it's there in case I need it.

I also re-built the project, but I'm not uploading it to LOL or to Itch.io.